MF-302 - Move things service mock from bootstrap to pkg.
